backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.2
